+++ huginn-proxy-lib/tests/backend/health_check/health.rs
@@ -1,20 +1,20 @@
-use huginn_proxy_lib::BackendHealth;
+use huginn_proxy_lib::UpstreamHealth;
 
 #[test]
 fn initial_state_is_healthy() {
-    let h = BackendHealth::new();
+    let h = UpstreamHealth::new();
     assert!(h.is_healthy());
 }
 
 #[test]
 fn default_is_healthy() {
-    let h = BackendHealth::default();
+    let h = UpstreamHealth::default();
     assert!(h.is_healthy());
 }
 
 #[test]
 fn set_unhealthy_and_recover() {
-    let h = BackendHealth::new();
+    let h = UpstreamHealth::new();
     h.set(false);
     assert!(!h.is_healthy());
     h.set(true);
@@ -23,7 +23,7 @@ fn set_unhealthy_and_recover() {
 
 #[test]
 fn idempotent_set() {
-    let h = BackendHealth::new();
+    let h = UpstreamHealth::new();
     h.set(true);
     h.set(true);
     assert!(h.is_healthy());
