+++ huginn-proxy-lib/tests/backend/mod.rs
@@ -1,2 +1,2 @@
 pub mod health_check;
-pub mod load_balancing;
+pub mod load_balance;
